+++ coordinator/points_writer_test.go
@@ -50,10 +50,89 @@ func TestPointsWriter_MapShards_One(t *testing.T) {
 	}
 }
 
+// Ensures the points writer maps to a new shard group when the shard duration
+// is changed.
+func TestPointsWriter_MapShards_AlterShardDuration(t *testing.T) {
+	ms := PointsWriterMetaClient{}
+	rp := NewRetentionPolicy("myp", time.Hour, 3)
+
+	ms.NodeIDFn = func() uint64 { return 1 }
+	ms.RetentionPolicyFn = func(db, retentionPolicy string) (*meta.RetentionPolicyInfo, error) {
+		return rp, nil
+	}
+
+	var (
+		i   int
+		now = time.Now()
+	)
+
+	ms.CreateShardGroupIfNotExistsFn = func(database, policy string, timestamp time.Time) (*meta.ShardGroupInfo, error) {
+		sg := []meta.ShardGroupInfo{
+			meta.ShardGroupInfo{
+				Shards:    make([]meta.ShardInfo, 1),
+				StartTime: now, EndTime: now.Add(rp.Duration).Add(-1),
+			},
+			meta.ShardGroupInfo{
+				Shards:    make([]meta.ShardInfo, 1),
+				StartTime: now.Add(time.Hour), EndTime: now.Add(3 * time.Hour).Add(rp.Duration).Add(-1),
+			},
+		}[i]
+		i++
+		return &sg, nil
+	}
+
+	c := coordinator.NewPointsWriter()
+	c.MetaClient = ms
+
+	pr := &coordinator.WritePointsRequest{
+		Database:        "mydb",
+		RetentionPolicy: "myrp",
+	}
+	pr.AddPoint("cpu", 1.0, now, nil)
+	pr.AddPoint("cpu", 2.0, now.Add(2*time.Second), nil)
+
+	var (
+		shardMappings *coordinator.ShardMapping
+		err           error
+	)
+	if shardMappings, err = c.MapShards(pr); err != nil {
+		t.Fatalf("unexpected an error: %v", err)
+	}
+
+	if got, exp := len(shardMappings.Points[0]), 2; got != exp {
+		t.Fatalf("got %d point(s), expected %d", got, exp)
+	}
+
+	if got, exp := len(shardMappings.Shards), 1; got != exp {
+		t.Errorf("got %d shard(s), expected %d", got, exp)
+	}
+
+	// Now we alter the retention policy duration.
+	rp.ShardGroupDuration = 3 * time.Hour
+
+	pr = &coordinator.WritePointsRequest{
+		Database:        "mydb",
+		RetentionPolicy: "myrp",
+	}
+	pr.AddPoint("cpu", 1.0, now.Add(2*time.Hour), nil)
+
+	// Point is beyond previous shard group so a new shard group should be
+	// created.
+	if shardMappings, err = c.MapShards(pr); err != nil {
+		t.Fatalf("unexpected an error: %v", err)
+	}
+
+	// We can check value of i since it's only incremeneted when a shard group
+	// is created.
+	if got, exp := i, 2; got != exp {
+		t.Fatal("new shard group was not created, expected it to be")
+	}
+}
+
 // Ensures the points writer maps a multiple points across shard group boundaries.
 func TestPointsWriter_MapShards_Multiple(t *testing.T) {
 	ms := PointsWriterMetaClient{}
-	rp := NewRetentionPolicy("myp", 0, 3)
+	rp := NewRetentionPolicy("myp", time.Hour, 3)
 	rp.ShardGroupDuration = time.Hour
 	AttachShardGroupInfo(rp, []meta.ShardOwner{
 		{NodeID: 1},
@@ -90,9 +169,9 @@ func TestPointsWriter_MapShards_Multiple(t *testing.T) {
 
 	// Three points that range over the shardGroup duration (1h) and should map to two
 	// distinct shards
-	pr.AddPoint("cpu", 1.0, time.Unix(0, 0), nil)
-	pr.AddPoint("cpu", 2.0, time.Unix(0, 0).Add(time.Hour), nil)
-	pr.AddPoint("cpu", 3.0, time.Unix(0, 0).Add(time.Hour+time.Second), nil)
+	pr.AddPoint("cpu", 1.0, time.Now(), nil)
+	pr.AddPoint("cpu", 2.0, time.Now().Add(time.Hour), nil)
+	pr.AddPoint("cpu", 3.0, time.Now().Add(time.Hour+time.Second), nil)
 
 	var (
 		shardMappings *coordinator.ShardMapping
@@ -107,12 +186,12 @@ func TestPointsWriter_MapShards_Multiple(t *testing.T) {
 	}
 
 	for _, points := range shardMappings.Points {
-		// First shard shoud have 1 point w/ first point added
+		// First shard should have 1 point w/ first point added
 		if len(points) == 1 && points[0].Time() != pr.Points[0].Time() {
 			t.Fatalf("MapShards() value mismatch. got %v, exp %v", points[0].Time(), pr.Points[0].Time())
 		}
 
-		// Second shard shoud have the last two points added
+		// Second shard should have the last two points added
 		if len(points) == 2 && points[0].Time() != pr.Points[1].Time() {
 			t.Fatalf("MapShards() value mismatch. got %v, exp %v", points[0].Time(), pr.Points[1].Time())
 		}
@@ -195,11 +274,15 @@ func TestPointsWriter_WritePoints(t *testing.T) {
 			RetentionPolicy: test.retentionPolicy,
 		}
 
+		// Ensure that the test shard groups are created before the points
+		// are created.
+		ms := NewPointsWriterMetaClient()
+
 		// Three points that range over the shardGroup duration (1h) and should map to two
 		// distinct shards
-		pr.AddPoint("cpu", 1.0, time.Unix(0, 0), nil)
-		pr.AddPoint("cpu", 2.0, time.Unix(0, 0).Add(time.Hour), nil)
-		pr.AddPoint("cpu", 3.0, time.Unix(0, 0).Add(time.Hour+time.Second), nil)
+		pr.AddPoint("cpu", 1.0, time.Now(), nil)
+		pr.AddPoint("cpu", 2.0, time.Now().Add(time.Hour), nil)
+		pr.AddPoint("cpu", 3.0, time.Now().Add(time.Hour+time.Second), nil)
 
 		// copy to prevent data race
 		theTest := test
@@ -245,7 +328,6 @@ func TestPointsWriter_WritePoints(t *testing.T) {
 			},
 		}
 
-		ms := NewPointsWriterMetaClient()
 		ms.DatabaseFn = func(database string) *meta.DatabaseInfo {
 			return nil
 		}
@@ -357,7 +439,7 @@ func TestBufferedPointsWriter(t *testing.T) {
 
 	numPoints := int(float64(capacity) * 5.5)
 	for i := 0; i < numPoints; i++ {
-		req.AddPoint("cpu", float64(i), time.Unix(0, 0).Add(time.Duration(i)*time.Second), nil)
+		req.AddPoint("cpu", float64(i), time.Now().Add(time.Duration(i)*time.Second), nil)
 	}
 
 	r := coordinator.IntoWriteRequest(req)
@@ -508,6 +590,7 @@ func NewRetentionPolicy(name string, duration time.Duration, nodeCount int) *met
 		Owners: owners,
 	})
 
+	start := time.Now()
 	rp := &meta.RetentionPolicyInfo{
 		Name:               "myrp",
 		ReplicaN:           nodeCount,
@@ -516,8 +599,8 @@ func NewRetentionPolicy(name string, duration time.Duration, nodeCount int) *met
 		ShardGroups: []meta.ShardGroupInfo{
 			meta.ShardGroupInfo{
 				ID:        nextShardID(),
-				StartTime: time.Unix(0, 0),
-				EndTime:   time.Unix(0, 0).Add(duration).Add(-1),
+				StartTime: start,
+				EndTime:   start.Add(duration).Add(-1),
 				Shards:    shards,
 			},
 		},
@@ -528,7 +611,7 @@ func NewRetentionPolicy(name string, duration time.Duration, nodeCount int) *met
 func AttachShardGroupInfo(rp *meta.RetentionPolicyInfo, owners []meta.ShardOwner) {
 	var startTime, endTime time.Time
 	if len(rp.ShardGroups) == 0 {
-		startTime = time.Unix(0, 0)
+		startTime = time.Now()
 	} else {
 		startTime = rp.ShardGroups[len(rp.ShardGroups)-1].StartTime.Add(rp.ShardGroupDuration)
 	}