+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## v1.0.2 [unreleased]
 
-## Bugfixes
+### Bugfixes
 
 - [#7150](https://github.com/influxdata/influxdb/issues/7150): Do not automatically reset the shard duration when using ALTER RETENTION POLICY
+- [#5878](https://github.com/influxdata/influxdb/issues/5878): Ensure correct shard groups created when retention policy has been altered.
+
 
 ## v1.0.1 [2016-09-26]